+++ docs/src/howto/firstgenrepeater_lowlevel/firstgenrepeater-04.swapcircuit.png
