+++ docs/src/howto/firstgenrepeater_lowlevel/firstgenrepeater-06.purcircuit2.png
