+++ examples/firstgenrepeater_lowlevel/5_clifford_full_example.jl
