+++ examples/firstgenrepeater_lowlevel/6.1_compare_formalisms_noplot.jl
