+++ examples/firstgenrepeater_lowlevel/6_compare_formalisms.jl
