+++ examples/firstgenrepeater_lowlevel/clifford_setup.jl
