--- test/examples/firstgenrepeater_4_tests.jl
@@ -1,5 +1,0 @@
-using Test
-
-@testset "Examples - firstgenrepeater 4" begin
-    include("../../examples/firstgenrepeater/4_visualization.jl")
-end
