--- test/examples/firstgenrepeater_5_tests.jl
@@ -1,5 +1,0 @@
-using Test
-
-@testset "Examples - firstgenrepeater 5" begin
-    include("../../examples/firstgenrepeater/5_clifford_full_example.jl")
-end
