--- test/examples/firstgenrepeater_6_1_tests.jl
@@ -1,5 +1,0 @@
-using Test
-
-@testset "Examples - firstgenrepeater 6.1" begin
-    include("../../examples/firstgenrepeater/6.1_compare_formalisms_noplot.jl")
-end
