--- test/examples/firstgenrepeater_6_tests.jl
@@ -1,5 +1,0 @@
-using Test
-
-@testset "Examples - firstgenrepeater 6" begin
-    include("../../examples/firstgenrepeater/6_compare_formalisms.jl")
-end
