+++ test/examples/firstgenrepeater_lowlevel_1_tests.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "Examples - firstgenrepeater_lowlevel 1" begin
+    include("../../examples/firstgenrepeater_lowlevel/1_entangler_example.jl")
+end
