+++ test/examples/firstgenrepeater_lowlevel_2_tests.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "Examples - firstgenrepeater_lowlevel 2" begin
+    include("../../examples/firstgenrepeater_lowlevel/2_swapper_example.jl")
+end
