+++ test/examples/firstgenrepeater_lowlevel_3_tests.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "Examples - firstgenrepeater_lowlevel 3" begin
+    include("../../examples/firstgenrepeater_lowlevel/3_purifier_example.jl")
+end
