+++ test/examples/firstgenrepeater_lowlevel_4_tests.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "Examples - firstgenrepeater_lowlevel 4" begin
+    include("../../examples/firstgenrepeater_lowlevel/4_visualization.jl")
+end
