+++ test/examples/firstgenrepeater_lowlevel_5_tests.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "Examples - firstgenrepeater_lowlevel 5" begin
+    include("../../examples/firstgenrepeater_lowlevel/5_clifford_full_example.jl")
+end
