+++ test/examples/firstgenrepeater_lowlevel_6_1_tests.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "Examples - firstgenrepeater_lowlevel 6.1" begin
+    include("../../examples/firstgenrepeater_lowlevel/6.1_compare_formalisms_noplot.jl")
+end
