+++ test/examples/firstgenrepeater_lowlevel_6_tests.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "Examples - firstgenrepeater_lowlevel 6" begin
+    include("../../examples/firstgenrepeater_lowlevel/6_compare_formalisms.jl")
+end
