--- test/examples/firstgenrepeater_protocolzoo_1_tests.jl
@@ -1,5 +1,0 @@
-using Test
-
-@testset "Examples - firstgenrepeater_protocolzoo 1" begin
-    include("../../examples/firstgenrepeater_protocolzoo/1_entangler_example.jl")
-end
