--- test/examples/firstgenrepeater_protocolzoo_3_tests.jl
@@ -1,5 +1,0 @@
-using Test
-
-@testset "Examples - firstgenrepeater_protocolzoo 3" begin
-    include("../../examples/firstgenrepeater_protocolzoo/3_purifier_example.jl")
-end
